created UlsTools module, migrated version check, created cli_installcheck
